New Change Approval System for non Admin users

+++ lib/providers/historic_sites_provider.dart
@@ -2,20 +2,25 @@ import 'dart:io' as io;
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
-import 'package:ringfort_app/models/user_data.dart';
 
-import '../helpers/location_helper.dart';
+import '../models/historic_site_staging.dart';
+import '../models/user_data.dart';
 import '../models/historic_site.dart';
 import '../firebase/firebaseDB.dart';
+import '../helpers/location_helper.dart';
 
 class HistoricSitesProvider with ChangeNotifier {
   final FirebaseDB firebaseDB = FirebaseDB();
 
   //-------------------------------------------------------------
-  // private list of ringfort sites and filtered
+  // private list of ringfort sites and filtered Sites
+  // Also a list of Staging changes awaiting approval
   //-------------------------------------------------------------
   List<HistoricSite> _sites = [];
   List<HistoricSite> _filteredSites = [];
+  List<HistoricSiteStaging> _stagingSites = [];
+  List<HistoricSiteStaging> _awaitingApprovalSites = [];
+  List<HistoricSiteStaging> _userApprovalHistory = [];
 
   //-------------------------------------------------------------
   // This is a getter for the private sites list which
@@ -27,6 +32,18 @@ class HistoricSitesProvider with ChangeNotifier {
 
   List<HistoricSite> get filteredSites {
     return [..._filteredSites];
+  }
+
+  List<HistoricSiteStaging> get stagingSites {
+    return [..._stagingSites];
+  }
+
+  List<HistoricSiteStaging> get awaitingApprovalSites {
+    return [..._awaitingApprovalSites];
+  }
+
+  List<HistoricSiteStaging> get userApprovalHistory {
+    return [..._userApprovalHistory];
   }
 
   //-------------------------------------------------------------
@@ -104,6 +121,45 @@ class HistoricSitesProvider with ChangeNotifier {
   }
 
   //-------------------------------------------------------------
+  // Returns the latest Staging Ringfort data from Firestore and update
+  // the local List and then notify comsumer it's changed
+  //-------------------------------------------------------------
+  Future<void> fetchAndSetStagingRingforts() async {
+    final List<HistoricSiteStaging> loadedSites = [];
+    var snapshot = await firebaseDB.fetchStagingSites();
+
+    final documents = snapshot.docs.map((docs) => docs.data()).toList();
+    documents.forEach((site) {
+      HistoricSiteStaging ringfort = HistoricSiteStaging.fromJson(site);
+      loadedSites.add(ringfort);
+    });
+    _stagingSites = loadedSites;
+    _awaitingApprovalSites =
+        _stagingSites.where((site) => site.actionStatus == 'awaiting').toList();
+
+    notifyListeners();
+  }
+
+  //-------------------------------------------------------------
+  // Returns the list of Approval requests for this user
+  //-------------------------------------------------------------
+  Future<void> fetchAndSetUserApprovalHistory(String userUid) async {
+    final List<HistoricSiteStaging> loadedSites = [];
+    var snapshot = await firebaseDB.fetchStagingSites();
+
+    final documents = snapshot.docs.map((docs) => docs.data()).toList();
+    documents.forEach((site) {
+      HistoricSiteStaging ringfort = HistoricSiteStaging.fromJson(site);
+      if (ringfort.actionedBy == userUid) {
+        loadedSites.add(ringfort);
+      }
+    });
+    _userApprovalHistory = loadedSites;
+
+    notifyListeners();
+  }
+
+  //-------------------------------------------------------------
   // Returns a single Ringfort Site by the uid.
   //-------------------------------------------------------------
   HistoricSite findSiteById(String uid) {
@@ -111,9 +167,18 @@ class HistoricSitesProvider with ChangeNotifier {
   }
 
   //-------------------------------------------------------------
-  // Add a new site to the List
+  // Returns a single Staging Ringfort Site by the uid.
   //-------------------------------------------------------------
-  void addSite(HistoricSite site, io.File image) async {
+  HistoricSiteStaging findStagingSiteById(String uid) {
+    return _stagingSites.firstWhere((site) => site.uid == uid);
+  }
+
+  //-------------------------------------------------------------
+  // Add a new site to the List. Depending if the user is an
+  // admin user or not, the new site will either be added to the
+  // live list of sites, or the staging sites list for approval
+  //-------------------------------------------------------------
+  void addSite(UserData userData, HistoricSite site, io.File image) async {
     // generate a document id for the new document on FB
     var docId = await firebaseDB.generateDocumentId();
     site.uid = docId;
@@ -125,7 +190,7 @@ class HistoricSitesProvider with ChangeNotifier {
     // adding the image to Firebase storage
     final imageUrl = await FirebaseDB().addImage(image, docId);
 
-    // waiting for the firebase storage o return a Url to store with site
+    // waiting for the firebase storage to return a Url to store with site
     final newSite = HistoricSite(
       uid: site.uid,
       siteName: site.siteName,
@@ -142,12 +207,27 @@ class HistoricSitesProvider with ChangeNotifier {
       createdBy: site.createdBy,
     );
 
-    // adding site to local list
-    _sites.add(newSite);
+    // Set up the site data for staging
+    final newStagingSite = HistoricSiteStaging(
+        uid: site.uid,
+        action: 'add',
+        actionDate: DateTime.now(),
+        actionStatus: 'awaiting',
+        actionedBy: userData.uid,
+        updatedSite: newSite);
+
+    if (userData.adminUser) {
+      // adding site to local site list, but only if admin user.
+      _sites.add(newSite);
+    } else {
+      // else add it to the local staging list
+      _stagingSites.add(newStagingSite);
+    }
+
     _filteredSites = [..._sites];
 
     // adding to Firebase Firestore
-    firebaseDB.addSite(newSite);
+    firebaseDB.addSite(userData.adminUser, newSite, newStagingSite);
     // Notify consumers of changes
     notifyListeners();
   }
@@ -155,8 +235,8 @@ class HistoricSitesProvider with ChangeNotifier {
   //-------------------------------------------------------------
   // This will find the site to be updated and update it.
   //-------------------------------------------------------------
-  void updateSite(
-      String siteUid, HistoricSite updatedSite, io.File image) async {
+  void updateSite(UserData userData, String siteUid, HistoricSite updatedSite,
+      io.File image) async {
     final siteIndex = _sites.indexWhere((site) => site.uid == siteUid);
 
     // Store the image in Firebase Storage
@@ -172,12 +252,31 @@ class HistoricSitesProvider with ChangeNotifier {
     updatedSite.county = addressMap['county'];
     updatedSite.province = addressMap['province'];
 
-    // Update the Rinfort object in the List
-    _sites[siteIndex] = updatedSite;
-    _filteredSites = [..._sites];
+    // Set up the site data for staging
+    final updateStagingSite = HistoricSiteStaging(
+        uid: siteUid,
+        action: 'update',
+        actionDate: DateTime.now(),
+        actionStatus: 'awaiting',
+        actionedBy: updatedSite.lastUpdatedBy,
+        updatedSite: updatedSite);
 
-    // Update the document on Firestore
-    firebaseDB.updateSite(updatedSite);
+    // Update the Rinfort object in the List if admin user
+    if (userData.adminUser) {
+      _sites[siteIndex] = updatedSite;
+      _filteredSites = [..._sites];
+      // Update the live document on Firestore
+      firebaseDB.updateSite(updatedSite);
+    } else {
+      // else add it to the local staging list
+      _stagingSites.add(updateStagingSite);
+      // add new staging document on Firestore for update
+      firebaseDB.addSite(false, null, updateStagingSite);
+      // update the list of awaiting staging records.
+      _awaitingApprovalSites = _stagingSites
+          .where((site) => site.actionStatus == 'awaiting')
+          .toList();
+    }
 
     // Notify comsumers of the data
     notifyListeners();
@@ -186,9 +285,90 @@ class HistoricSitesProvider with ChangeNotifier {
   //---------------------------------------------------------------------
   // This will delete the site which matches the uid
   //---------------------------------------------------------------------
-  void deleteSite(String uid) {
-    _sites.removeWhere((site) => site.uid == uid);
-    firebaseDB.deleteSite(uid);
+  void deleteSite(UserData userData, HistoricSite deletSite) {
+    if (userData.adminUser) {
+      _sites.removeWhere((site) => site.uid == deletSite.uid);
+      firebaseDB.deleteSite(deletSite.uid);
+    } else {
+      // Set up the site data for staging
+      final deleteStagingSite = HistoricSiteStaging(
+          uid: deletSite.uid,
+          action: 'delete',
+          actionDate: DateTime.now(),
+          actionStatus: 'awaiting',
+          actionedBy: userData.uid,
+          updatedSite: deletSite);
+
+      // else add it to the local staging list
+      _stagingSites.add(deleteStagingSite);
+      // add new staging document on Firestore for update
+      firebaseDB.addSite(false, null, deleteStagingSite);
+      // update the list of awaiting staging records.
+      _awaitingApprovalSites = _stagingSites
+          .where((site) => site.actionStatus == 'awaiting')
+          .toList();
+    }
+    notifyListeners();
+  }
+
+  //----------------------------------------------------------------------
+  //  This will approve the staging site. Adding it to the historicSites
+  // collection and updating on the historicSitesStaging to approved
+  //----------------------------------------------------------------------
+  void approveStagingSite(String uid) {
+    // Find the staging site object
+    var historicSiteStaging =
+        _stagingSites.firstWhere((site) => site.uid == uid);
+
+    // Add to local sites
+    if (historicSiteStaging.action == 'add') {
+      // Add to local sites
+      _sites.add(historicSiteStaging.updatedSite);
+      // Adding to Firestore historicSites collection
+      firebaseDB.addSite(true, historicSiteStaging.updatedSite, null);
+    } else if (historicSiteStaging.action == 'update') {
+      // update in local site list
+      _sites[_sites.indexWhere((site) => site.uid == uid)] =
+          historicSiteStaging.updatedSite;
+      // updating in the firetone historicSites collection
+      firebaseDB.updateSite(historicSiteStaging.updatedSite);
+    } else {
+      // delete from local site list
+      _sites.removeWhere((site) => site.uid == uid);
+      // delete from the Firestore historicSites collection
+      firebaseDB.deleteSite(uid);
+    }
+
+    // update to approved in the local staging site list
+    historicSiteStaging.actionStatus = 'approved';
+
+    // update the list of awaiting staging records.
+    _awaitingApprovalSites =
+        _stagingSites.where((site) => site.actionStatus == 'awaiting').toList();
+
+    // Update from the historicSitesStaging firestore collection
+    firebaseDB.updateStagingSite(historicSiteStaging);
+
+    notifyListeners();
+  }
+
+  //----------------------------------------------------------------------
+  //  This will reject the staging site. Updating on the
+  //  historicSitesStaging status to rejected
+  //----------------------------------------------------------------------
+  void rejectStagingSite(String uid) {
+    // Find the staging site object
+    var historicSiteStaging =
+        _stagingSites.firstWhere((site) => site.uid == uid);
+
+    // update to rejected in the local staging site list
+    historicSiteStaging.actionStatus = 'rejected';
+    _awaitingApprovalSites =
+        _stagingSites.where((site) => site.actionStatus == 'awaiting').toList();
+
+    // Update from the historicSitesStaging firestore collection
+    firebaseDB.updateStagingSite(historicSiteStaging);
+
     notifyListeners();
   }
 }
